+++ web/app/resources/page.tsx
@@ -1,0 +1,5 @@
+import ResourcesModule from '../../components/resources/ResourcesModule';
+
+export default function ResourcesPage() {
+  return <ResourcesModule />;
+}